Fixed 64-bit issues in COM object and samples/tests (XERCESC-1531)
git-svn-id: https://svn.apache.org/repos/asf/xerces/c/trunk@556537 13f79535-47bb-0310-9956-ffa450edef68

+++ src/xercesc/com/IXMLDOMCharacterDataImpl.h
@@ -85,7 +85,7 @@ STDMETHOD(get_length)(long  *pVal)
 
 	try
 	{
-		*pVal = get_DOMCharacterData()->getLength();
+		*pVal = (long)get_DOMCharacterData()->getLength();
 	}
 	catch(...)
 	{
